String resolving concept
The struct _strings encapsulates two string tables, CTF and ELF.
A pointer to this struct is handed to labels, types and other reading
functions, as it is used to resolve the name references. After all is
resolved, the contents are freed, as all the reading functions make a
duplicate string.