Grounds for the three-pass type parsing
All these unions where introduced with one simple goal: to allow forward
references in the type ID references. Therefore, in the first pass of the
parsing algorithm we store only the integer ID, whereas in the third pass we
solve all this references and we do not need the ID anymore. Therefore, these
two information are mutually exclusive in use and we can have them in an union,
thus waisting no space.